chore(List): refactor to prevent breaking
