+++ packages/components/src/List/ListComposition/constants.js
@@ -1,3 +1,8 @@
+export const DISPLAY_MODE = {
+	TABLE: 'table',
+	LARGE: 'large',
+};
+
 export const SORT = {
 	ASC: 'ASC',
 	DESC: 'DESC',
