--- packages/components/src/List/Toolbar/constants.js
@@ -1,4 +1,0 @@
-export const DISPLAY_MODE = {
-	TABLE: 'table',
-	LARGE: 'large',
-};
